requirements.txt
tensorflow==2.15
streamlit
pillow
numpy
scikit-learn
